refactor analyzer
